feat: update getting started page with new content
